+++ samples/gltf_viewer.cpp
@@ -346,7 +346,7 @@ static void createGroundPlane(Engine* engine, Scene* scene, App& app) {
 static float sGlobalScale = 1.0f;
 static float sGlobalScaleAnamorphism = 0.0f;
 static int sGlobalScaleQuality = 0;
-static float sGlobalScaleSharpness = 0.2f;
+static float sGlobalScaleSharpness = 0.9f;
 
 int main(int argc, char** argv) {
     App app;
@@ -578,7 +578,7 @@ int main(int argc, char** argv) {
                 ImGui::SliderFloat("scale", &sGlobalScale, 0.25f, 1.0f);
                 ImGui::SliderFloat("anamorphism", &sGlobalScaleAnamorphism, -1.0f, 1.0f);
                 ImGui::SliderInt("quality", &sGlobalScaleQuality, 0, 3);
-                ImGui::SliderFloat("sharpness", &sGlobalScaleSharpness, 0.0f, 2.0f);
+                ImGui::SliderFloat("sharpness", &sGlobalScaleSharpness, 0.0f, 1.0f);
 
             }